requirements.txt
matplotlib~=3.10.5
numpy~=2.3.2
Pillow~=11.3.0
PyYAML~=6.0.2
scikit_learn~=1.7.1
seaborn~=0.13.2
torch~=2.8.0
torchvision~=0.23.0
